requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7
bcrypt==4.0.1
python-dotenv==1.0.0
bleach==6.0.0
flask-talisman==1.1.0